# server/swagger.yaml
openapi: 3.0.3
info:
  title: MyLoadTest Datapool API
  description: The MyLoadTest Datapool is a centralised, persistent data store. ...TODO
  version: 0.0.2
  contact:
    name: MyLoadTest
    url: https://www.myloadtest.com
servers:
  - url: http://localhost:3000
    description: Self-hosted datapool
...
paths:
...
  /code/{module}:
    post:
      summary: Upload a function module
      description: "Upload a single Node.js module with named exports that may be called with ```POST /code/{module}/{function}``` or ```GET /code/{module}/{function}```. Note that the Swagger-UI page does not support file uploads properly, so files may be uploaded with curl e.g. ```curl -X POST 'http://localhost:3000/code/example' -H 'content-type: text/plain' --data-binary '@example.js'```. This is a massive security problem, so it might be best to disable this feature and manually copy user-code modules to the datapool server."
      deprecated: true
      tags: [function]
      parameters:
        - in: path
          name: module
          description: The name of the module
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          text/plain: # this is a file upload. See: https://swagger.io/docs/specification/describing-request-body/file-upload/
            schema:
              type: string
              format: binary
      responses:
        201:
          description: An empty response body is returned when the module has been uploaded successfully.
        400:
          description: There is something wrong with the uploaded module.
        409:
          description: Conflict. Module already exists.
    get:
      summary: Get a list of functions in the module
      description: Returns a list of named exports from the Node.js module. Any of these may be called as a function. Function parameters (argument lists) are not returned.
      tags: [function]
      parameters:
        - in: path
          name: module
          description: The name of the module
          required: true
          schema:
            type: string
      responses:
        200:
          description: A list of module names, or an empty list if no modules are available
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                example: ['generateAuthHeader', 'base64encode', 'base64decode']
        400:
          description: The module name is invalid
        404:
          description: The module was not found
    delete:
      summary: Remove a function module
      description: Delete a function module from the filesystem and clear the module cache.
      tags: [function]
      parameters:
        - in: path
          name: module
          description: The name of the module
          required: true
          schema:
            type: string
      responses:
        200:
          description: The module was deleted successfully
        400:
          description: The module name is invalid
        404:
          description: The module was not found
  /code/{module}/{function}:
    post:
      summary: Call a function with arguments
      description: Invoke a function with an array of arguments (in the POST body) passed to the function.
      tags: [function]
      parameters:
        - in: path
          name: module
          description: The name of the module
          required: true
          schema:
            type: string
        - in: path
          name: function
          description: The name of the function
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: Function arguments in array may be any type
              type: array
              items:
                anyOf:
                  - type: string
                  - type: number
                  - type: object
              example: ['hello', 'world']
      responses:
        201:
          description: An empty response body is returned when the counter is initialised or reset.
        400:
          description: Bad request. Invalid module/function name, or invalid POST body
        404:
          description: The module or the function was not found
        500:
          description: The function threw a runtime exception
    get:
      summary: Call a function with no arguments
      description: Invoke a function with no arguments passed to the function
      tags: [function]
      parameters:
        - in: path
          name: module
          description: The name of the module
          required: true
          schema:
            type: string
        - in: path
          name: function
          description: The name of the function
          required: true
          schema:
            type: string
      responses:
        200:
          description: The function's return value in JSON format
        400:
          description: The module or function name is invalid
        404:
          description: The module or function was not found
        500:
          description: The function threw a runtime exception